requirements.txt
aiokafka==0.12.0
alembic==1.15.2
asyncpg==0.30.0
black==25.1.0
celery==5.4.0
fastapi[all]==0.115.12
kafka-python==2.1.4
passlib==1.7.4
pyjwt==2.10.1
redis==5.2.1
sqlalchemy==2.0.40
uvicorn==0.34.0
